+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
+- Initial release
+- Added multi-ring circular progress widget
+- Supports animations and custom center widget
 
-* TODO: Describe initial release.